python/requirements-dev.txt
# Development dependencies for FurSight Python SDK

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
requests-mock>=1.10.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Type stubs
types-requests>=2.28.0
types-setuptools>=65.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0
myst-parser>=0.18.0

# Build tools
build>=0.10.0
twine>=4.0.0

# Security
bandit>=1.7.0
safety>=2.3.0

# Development utilities
pre-commit>=3.0.0
tox>=4.0.0
